+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - **git-ignore:** ignore sitemap generated items
 - **license:** added license
 - **project-list-card:** update appearance to be nicer
+- **project-list-card:** general improvement
 - **rustywind:** added rustywind to dev dependencies
 - **sharp:** added sharp for nextjs
 - **sidebar:** added sidebar to whole site